SFA: Move shader code to its own file
