+++ src/StarFoxAdventures/models.ts
@@ -11,88 +11,6 @@ import { GfxRenderInstManager } from "../gfx/render/GfxRenderer";
 
 import { SFATexture } from './textures';
 
-function dataSubarray(data: DataView, byteOffset: number, byteLength?: number): DataView {
-    return new DataView(data.buffer, data.byteOffset + byteOffset, byteLength);
-}
-
-interface ShaderLayer {
-    texNum: number;
-    tevMode: number;
-}
-
-export interface Shader {
-    layers: ShaderLayer[],
-    enableCull: boolean;
-    flags: number;
-    hasTexmtx01: boolean;
-    hasTexmtx2: boolean;
-    attrFlags: number;
-}
-
-function parseShaderLayer(data: DataView): ShaderLayer {
-    return {
-        texNum: data.getUint32(0),
-        tevMode: data.getUint8(4),
-    };
-}
-
-interface ShaderFields {
-    size: number;
-    numLayers: number;
-    layers: number;
-}
-
-export const SFA_SHADER_FIELDS: ShaderFields = {
-    size: 0x44,
-    numLayers: 0x41,
-    layers: 0x24,
-};
-
-export const EARLY_SFA_SHADER_FIELDS: ShaderFields = {
-    size: 0x40,
-    numLayers: 0x3b,
-    layers: 0x24, // ???
-};
-
-enum ShaderFlags {
-    Cull = 0x8,
-}
-
-export function parseShader(data: DataView, fields: ShaderFields): Shader {
-    const shader: Shader = {
-        layers: [],
-        enableCull: false,
-        flags: 0,
-        hasTexmtx01: false,
-        hasTexmtx2: false,
-        attrFlags: 0,
-    };
-
-    let numLayers = data.getUint8(fields.numLayers);
-    if (numLayers > 2) {
-        console.warn(`Number of shader layers greater than maximum (${numLayers} / 2)`);
-        numLayers = 2;
-    }
-    for (let i = 0; i < numLayers; i++) {
-        const layer = parseShaderLayer(dataSubarray(data, fields.layers + i * 8));
-        shader.layers.push(layer);
-    }
-
-    shader.flags = data.getUint32(0x3c);
-    // FIXME: find this field's offset for demo files
-    shader.enableCull = (shader.flags & ShaderFlags.Cull) != 0;
-
-    // FIXME: the texmtx stuff below is broken or not present in SFA...
-    // shader.hasTexmtx01 = data.getUint32(offs + 8) == 1 || data.getUint32(offs + 20) == 1;
-    // shader.hasTexmtx2 = (data.getUint32(offs + 64 + 2) & 0x80) != 0;
-    shader.hasTexmtx01 = data.getUint32(0x34) != 0;
-    shader.hasTexmtx2 = false;
-
-    shader.attrFlags = data.getUint8(0x40);
-
-    return shader
-}
-
 export class ModelInstance {
     private loadedVertexLayout: LoadedVertexLayout;
     private loadedVertexData: LoadedVertexData;
